Fixed bug with winrate

+++ PlayerStats/Player.cs
@@ -25,7 +25,7 @@ using System;
              get
              {
                  if(playedGames>0)
-                     return wonGames/playedGames;
+                     return (float)wonGames/(float)playedGames;
                  else 
                      return 0;
              }
